ADM-365 fix styles
