+++ frontend/src/components/Metrics/MetricsStep/DeploymentFrequencySettings/SingleSelection/style.tsx
@@ -2,6 +2,7 @@ import { styled } from '@mui/material/styles'
 import { FormControl } from '@mui/material'
 
 export const FormControlWrapper = styled(FormControl)({
-  marginBottom: '1rem',
+  margin: '1.5rem 0',
   width: '100%',
+  height: '2rem',
 })
